Move tests to collection_length and add tests for Literal and non-limited List
